Correções de segurança na category

+++ ToDoFlow.Backend/ToDoFlow.Services/Services/CategoryService.cs
@@ -12,14 +12,15 @@ namespace ToDoFlow.Services.Services
         private readonly ICategoryRepository _categoryRepository = categoryRepository;
         private readonly IMapper _mapper = mapper;
 
-        public async Task<ApiResponse<List<CategoryReadDto>>> CreateCategoryAsync(CategoryCreateDto categoryCreateDto)
+        public async Task<ApiResponse<List<CategoryReadDto>>> CreateCategoryAsync(int userId, CategoryCreateDto categoryCreateDto)
         {
             try
             {
                 Category category = _mapper.Map<Category>(categoryCreateDto);
-                await _categoryRepository.CreateCategoryAsync(category);
+                category.UserId = userId;
 
-                List<Category> categories = await _categoryRepository.ReadCategoryByUserAsync(category.UserId);
+                List<Category> categories = await _categoryRepository.CreateCategoryAsync(category);
+
                 List<CategoryReadDto> categoryReadDtos = _mapper.Map<List<CategoryReadDto>>(categories);
 
                 return new ApiResponse<List<CategoryReadDto>>(categoryReadDtos, true, "Category created successfully", 201);
@@ -36,6 +37,7 @@ namespace ToDoFlow.Services.Services
             {
                 List<Category> categories = await _categoryRepository.ReadCategoryAsync();
                 List<CategoryReadDto> categoryReadDtos = _mapper.Map<List<CategoryReadDto>>(categories);
+
 
                 return new ApiResponse<List<CategoryReadDto>>(categoryReadDtos, true, "Operation carried out successfully", 200);
             }
@@ -60,11 +62,22 @@ namespace ToDoFlow.Services.Services
             }
         }
 
-        public async Task<ApiResponse<CategoryReadDto>> ReadCategoryByIdAsync(int id)
+        public async Task<ApiResponse<CategoryReadDto>> ReadCategoryByIdAsync(int id, int userId)
         {
             try
             {
                 Category category = await _categoryRepository.ReadCategoryByIdAsync(id);
+                
+                if (category == null)
+                {
+                    return new ApiResponse<CategoryReadDto>(null, false, "Erro: Category not found", 404);
+                }
+
+                if (category.UserId != userId)
+                {
+                    return new ApiResponse<CategoryReadDto>(null, false, "Unauthorized", 403);
+                }
+
                 CategoryReadDto categoryReadDto = _mapper.Map<CategoryReadDto>(category);
 
                 return new ApiResponse<CategoryReadDto>(categoryReadDto, true, "Operation carried out successfully", 200);
@@ -75,11 +88,22 @@ namespace ToDoFlow.Services.Services
             }
         }
 
-        public async Task<ApiResponse<List<CategoryReadDto>>> UpdateCategoryAsync(int id, CategoryUpdateDto categoryUpdateDto)
+        public async Task<ApiResponse<List<CategoryReadDto>>> UpdateCategoryAsync(int id, int userId, CategoryUpdateDto categoryUpdateDto)
         {
             try
             {
                 Category category = await _categoryRepository.ReadCategoryByIdAsync(id);
+
+                if (category == null)
+                {
+                    return new ApiResponse<List<CategoryReadDto>>(null, false, "Category not found", 404);
+                }
+
+                if (category.UserId != userId)
+                {
+                    return new ApiResponse<List<CategoryReadDto>>(null, false, "Unauthorized", 403);
+                }
+                
                 _mapper.Map(categoryUpdateDto, category);
                 await _categoryRepository.UpdateCategoryAsync(category);
 
@@ -94,10 +118,22 @@ namespace ToDoFlow.Services.Services
             }
         }
 
-        public async Task<ApiResponse<List<CategoryReadDto>>> DeleteCategoryAsync(int id)
+        public async Task<ApiResponse<List<CategoryReadDto>>> DeleteCategoryAsync(int id, int userId)
         {
             try
             {
+                Category category = await _categoryRepository.ReadCategoryByIdAsync(id);
+
+                if (category == null)
+                {
+                    return new ApiResponse<List<CategoryReadDto>>(null, false, "Category not found", 404);
+                }
+
+                if (category.UserId != userId)
+                {
+                    return new ApiResponse<List<CategoryReadDto>>(null, false, "Unauthorized", 403);
+                }
+
                 List<Category> categories = await _categoryRepository.DeleteCategoryAsync(id);
                 List<CategoryReadDto> categoryReadDtos = _mapper.Map<List<CategoryReadDto>>(categories);
